+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Rojo Changelog
 
 ## Unreleased Changes
+* Updated Undo/Redo history to be more robust ([#915])
 * Fixed removing trailing newlines ([#903])
 * Added Never option to Confirmation ([#893])
 * Added popout diff visualizer for table properties like Attributes and Tags ([#834])
@@ -52,7 +53,7 @@
  	| `rbxm`         | `.rbxm`         |
  	| `rbxmx`        | `.rbxmx`        |
  	| `project`      | `.project.json` |
-    | `yaml`         | `.yml`          |
+  | `yaml`         | `.yml`          |
  	| `ignore`       | None!           |
 
     **All** sync rules are reset between project files, so they must be specified in each one when nesting them. This is to ensure that nothing can break other projects by changing how files are synced!
@@ -69,6 +70,7 @@
 [#903]: https://github.com/rojo-rbx/rojo/pull/903
 [#911]: https://github.com/rojo-rbx/rojo/pull/911
 [#913]: https://github.com/rojo-rbx/rojo/pull/913
+[#915]: https://github.com/rojo-rbx/rojo/pull/915
 
 ## [7.4.1] - February 20, 2024
 * Made the `name` field optional on project files ([#870])